requirements.txt
asgiref==3.5.0
distlib==0.3.4
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djongo==1.3.6
dnspython==2.2.1
filelock==3.6.0
Pillow==9.1.0
platformdirs==2.5.1
PyJWT==2.3.0
pymongo==4.1.1
pytz==2022.1
six==1.16.0
sqlparse==0.2.4
tzdata==2022.1
virtualenv==20.14.0
virtualenvwrapper-win==1.2.7
